RYA-394 Lowercase org.apache.rya.indexing.IndexPlanValidator. Closes #320
