+++ extras/indexing/src/test/java/org/apache/rya/indexing/indexPlanValidator/IndexPlanValidatorTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.rya.indexing.IndexPlanValidator;
+package org.apache.rya.indexing.indexPlanValidator;
 
 import java.util.ArrayList;
 import java.util.Collection;
